% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violets.R
\name{violets}
\alias{violets}
\title{violets takes a model and tries to estimate it using OLS instead}
\usage{
violets(model, ...)
}
\arguments{
\item{model}{a fitted model object}
}
\description{
violets takes a model and tries to estimate it using OLS instead
}
\details{
\code{violets} tries to extract information from a model object using the
\code{formula} and \code{model.frame} methods. It will not work on objects that do not
conform to this standard.
}
